refactor: extract resolveTemplate() from <TemplateLoader />
