+++ packages/headless/src/components/TemplateLoader.tsx
@@ -1,49 +1,24 @@
 import React from 'react';
+import type { GetServerSidePropsContext, GetStaticPropsContext } from 'next';
+import type { ApolloClient, NormalizedCacheObject } from '@apollo/client';
 import { useNextUriInfo } from '../api';
-import { UriInfo } from '../types';
+import { resolveTemplate } from '../utils/resolveTemplate';
 
-function resolveTemplate(
-  pageInfo: UriInfo | undefined,
-): React.FunctionComponent | null {
-  /**
-   * List out files in main project using Webpack.
-   */
-  let context;
-
-  try {
-    context = require.context('theme', false, /.*/);
-  } catch (e) {
-    console.warn(
-      '"theme" directory not detected in Next.js project. Please add it to take advantage of <TemplateLoader />.',
-    );
-    return null;
-  }
-
-  if (!pageInfo || !pageInfo.templates) {
-    return context(`./index`).default as React.FunctionComponent;
-  }
-
-  // eslint-disable-next-line no-restricted-syntax
-  for (const template of pageInfo.templates) {
-    try {
-      const importedTemplate = context(`./${template}`);
-
-      if (importedTemplate) {
-        return importedTemplate.default as React.FunctionComponent;
-      }
-    } catch (e) {
-      // eslint-disable-next-line no-console
-      console.debug('Template not found.', e);
-    }
-  }
-
-  return context(`./index`).default as React.FunctionComponent;
+export interface Template {
+  default: React.FunctionComponent | null;
+  getPropsMiddleware?: (
+    /* eslint-disable @typescript-eslint/no-explicit-any */
+    promises: Array<Promise<unknown> | undefined>,
+    apolloClient: ApolloClient<NormalizedCacheObject>,
+    currentUrlPath: string,
+    context: GetStaticPropsContext | GetServerSidePropsContext,
+  ) => Array<Promise<unknown> | undefined>;
 }
 
 export default function TemplateLoader(): JSX.Element | null {
   const pageInfo = useNextUriInfo();
 
-  const Component = resolveTemplate(pageInfo);
+  const Component = resolveTemplate(pageInfo)?.default;
 
   if (!Component) {
     return null;
